Cleaned up db.js
